+++ backend-server/core/business/admin/service/role_service.go
@@ -3,7 +3,7 @@ package service
 import "github.com/vamika-digital/wms-api-server/core/business/admin/dto/role"
 
 type RoleService interface {
-	GetAllRoles(page int64, pageSize int64, sort string, filter *role.RoleFilterDto) ([]*role.RoleDto, int64, error)
+	GetAllRoles(page int16, pageSize int16, sort string, filter *role.RoleFilterDto) ([]*role.RoleDto, int64, error)
 	CreateRole(roleDto *role.RoleCreateDto) error
 	GetRoleByID(roleID int64) (*role.RoleDto, error)
 	UpdateRole(roleID int64, role *role.RoleUpdateDto) error
