+++ backend-server/core/business/master/service/customer_service.go
@@ -3,7 +3,7 @@ package service
 import "github.com/vamika-digital/wms-api-server/core/business/master/dto/customer"
 
 type CustomerService interface {
-	GetAllCustomers(page int64, pageSize int64, sort string, filter *customer.CustomerFilterDto) ([]*customer.CustomerDto, int64, error)
+	GetAllCustomers(page int16, pageSize int16, sort string, filter *customer.CustomerFilterDto) ([]*customer.CustomerDto, int64, error)
 	CreateCustomer(customerDto *customer.CustomerCreateDto) error
 	GetCustomerByID(customerID int64) (*customer.CustomerDto, error)
 	UpdateCustomer(customerID int64, customer *customer.CustomerUpdateDto) error
