+++ backend-server/core/business/master/service/machine_service.go
@@ -3,7 +3,7 @@ package service
 import "github.com/vamika-digital/wms-api-server/core/business/master/dto/machine"
 
 type MachineService interface {
-	GetAllMachines(page int64, pageSize int64, sort string, filter *machine.MachineFilterDto) ([]*machine.MachineDto, int64, error)
+	GetAllMachines(page int16, pageSize int16, sort string, filter *machine.MachineFilterDto) ([]*machine.MachineDto, int64, error)
 	CreateMachine(machineDto *machine.MachineCreateDto) error
 	GetMachineByID(machineID int64) (*machine.MachineDto, error)
 	GetMachineByCode(machineCode string) (*machine.MachineDto, error)
