+++ backend-server/core/business/master/service/store_service.go
@@ -3,7 +3,7 @@ package service
 import "github.com/vamika-digital/wms-api-server/core/business/master/dto/store"
 
 type StoreService interface {
-	GetAllStores(page int64, pageSize int64, sort string, filter *store.StoreFilterDto) ([]*store.StoreDto, int64, error)
+	GetAllStores(page int16, pageSize int16, sort string, filter *store.StoreFilterDto) ([]*store.StoreDto, int64, error)
 	CreateStore(storeDto *store.StoreCreateDto) error
 	GetStoreByID(storeID int64) (*store.StoreDto, error)
 	GetStoreByCode(storeCode string) (*store.StoreDto, error)
